Add further coverage tests.
FossilOrigin-Name: 1fd1d03d347729f83e71c9b4f0cf09f6421e576cc38f39bbed361c0396f50c59

+++ src/hct_file.c
@@ -1247,7 +1247,7 @@ static u64 hctFilePageToChunk(HctMapping *pMapping, i64 iPg){
 ** object. Do not create a new mapping.
 */
 static int hctFileUpdateMappingForSlot(HctFile *pFile, i64 iSlot){
-  int iChunk = hctFilePageToChunk(pFile->pMapping, iSlot);
+  u64 iChunk = hctFilePageToChunk(pFile->pMapping, iSlot);
   if( iChunk>=pFile->pMapping->nChunk ){
     HctMapping *pOld = pFile->pMapping;
     hctFileEnterServerMutex(pFile);
